alter container-id determination for compatibility
previously, we assumed that all cgroups within a container would be
in the form /docker/<container id>. While this is usually the case, it
is not the case on all systems. In particular, it was seen that "rdma"
was running as the root cgroup, when viewed from within a container.

With that in mind, the best heuristic instead seems to be to look for
*any* cgroup of the form /docker/<container id>, and taking the
container id from that, if any such cgroup is found.